# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisEnvelope": {
        "description": "Provenance: describes the served object, never alters it.",
        "properties": {
          "analysis_methodology_hash": {
            "title": "Analysis Methodology Hash",
            "type": "string"
          },
          "analysis_version": {
            "title": "Analysis Version",
            "type": "string"
          },
          "data_as_of": {
            "format": "date",
            "title": "Data As Of",
            "type": "string"
          },
          "document_sha256": {
            "title": "Document Sha256",
            "type": "string"
          },
          "meta_version": {
            "title": "Meta Version",
            "type": "string"
          },
          "methodology_hash": {
            "title": "Methodology Hash",
            "type": "string"
          },
          "sections": {
            "items": {
              "type": "string"
            },
            "title": "Sections",
            "type": "array"
          },
          "security_id": {
            "title": "Security Id",
            "type": "string"
          },
          "snapshot_generated_at": {
            "title": "Snapshot Generated At",
            "type": "string"
          },
          "weekly_version": {
            "title": "Weekly Version",
            "type": "string"
          }
        },
        "required": [
          "security_id",
          "meta_version",
          "snapshot_generated_at",
          "data_as_of",
          "weekly_version",
          "methodology_hash",
          "analysis_version",
          "analysis_methodology_hash",
          "document_sha256",
          "sections"
        ],
        "title": "AnalysisEnvelope",
        "type": "object"
      },
      "AnalysisResponse": {
        "properties": {
          "document": {
            "additionalProperties": true,
            "title": "Document",
            "type": "object"
          },
          "envelope": {
            "$ref": "#/components/schemas/AnalysisEnvelope"
          }
        },
        "required": [
          "envelope",
          "document"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
...
      "EventsEnvelope": {
        "properties": {
          "analysis_version": {
            "title": "Analysis Version",
            "type": "string"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "data_as_of": {
            "format": "date",
            "title": "Data As Of",
            "type": "string"
          },
          "dataset": {
            "title": "Dataset",
            "type": "string"
          },
          "meta_version": {
            "title": "Meta Version",
            "type": "string"
          },
          "physical_sha256": {
            "title": "Physical Sha256",
            "type": "string"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "security_id": {
            "title": "Security Id",
            "type": "string"
          },
          "snapshot_generated_at": {
            "title": "Snapshot Generated At",
            "type": "string"
          }
        },
        "required": [
          "security_id",
          "meta_version",
          "snapshot_generated_at",
          "data_as_of",
          "analysis_version",
          "dataset",
          "content_sha256",
          "physical_sha256",
          "row_count"
        ],
        "title": "EventsEnvelope",
        "type": "object"
      },
      "EventsResponse": {
        "properties": {
          "envelope": {
            "$ref": "#/components/schemas/EventsEnvelope"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "envelope",
          "rows",
          "next_cursor"
        ],
        "title": "EventsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "data_as_of": {
            "format": "date",
            "title": "Data As Of",
            "type": "string"
          },
          "meta_version": {
            "title": "Meta Version",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SecuritySummary"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "meta_version",
          "data_as_of",
          "query",
          "results"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "SecuritySummary": {
        "properties": {
          "analytical_universe": {
            "title": "Analytical Universe",
            "type": "boolean"
          },
          "instrument_type": {
            "title": "Instrument Type",
            "type": "string"
          },
          "isin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isin"
          },
          "last_date": {
            "format": "date",
            "title": "Last Date",
            "type": "string"
          },
          "listing_status": {
            "title": "Listing Status",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "security_id": {
            "title": "Security Id",
            "type": "string"
          },
          "series": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "usable_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usable From"
          }
        },
        "required": [
          "security_id",
          "symbol",
          "name",
          "isin",
          "series",
          "listing_status",
          "instrument_type",
          "analytical_universe",
          "status",
          "usable_from",
          "last_date"
        ],
        "title": "SecuritySummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "See the structure. Read the trend.",
    "title": "ChartLens API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/securities": {
      "get": {
        "description": "Search current and past symbols, ISINs and names.",
        "operationId": "search_api_v1_securities_get",
        "parameters": [
          {
            "description": "Symbol, ISIN or name",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Symbol, ISIN or name",
              "maxLength": 40,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "universe",
            "required": false,
            "schema": {
              "default": "analytical",
              "enum": [
                "analytical",
                "all"
              ],
              "title": "Universe",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "securities"
        ]
      }
    },
...
    "/api/v1/securities/{security_id}/analysis": {
      "get": {
        "description": "The published analysis of one security, verbatim, or whole named sections of it.",
        "operationId": "analysis_api_v1_securities__security_id__analysis_get",
        "parameters": [
          {
            "in": "path",
            "name": "security_id",
            "required": true,
            "schema": {
              "title": "Security Id",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated top-level sections; default: the whole document",
            "in": "query",
            "name": "sections",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated top-level sections; default: the whole document",
              "title": "Sections"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "unknown_security, not_analysed, no_analysis_in_snapshot"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analysis",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/securities/{security_id}/breakout-events": {
      "get": {
        "description": "Rows of one published breakout-event dataset matching explicit predicates on\nstored fields, in the stored order. Retrieval, never ranking or selection.",
        "operationId": "breakout_events_api_v1_securities__security_id__breakout_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "security_id",
            "required": true,
            "schema": {
              "title": "Security Id",
              "type": "string"
            }
          },
          {
            "description": "One dataset; never merged",
            "in": "query",
            "name": "source",
            "required": true,
            "schema": {
              "description": "One dataset; never merged",
              "enum": [
                "pattern",
                "level"
              ],
              "title": "Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "BREAKOUT",
                    "BREAKDOWN"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Direction"
            }
          },
          {
            "description": "source=pattern only",
            "in": "query",
            "name": "pattern_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "source=pattern only",
              "title": "Pattern Type"
            }
          },
          {
            "description": "source=level only",
            "in": "query",
            "name": "level_source_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "source=level only",
              "title": "Level Source Type"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "unknown_security, not_analysed, no_analysis_in_snapshot"
          },
          "409": {
            "description": "the snapshot changed between pages; restart without a cursor"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Breakout Events",
        "tags": [
          "analysis"
        ]
      }
    },
...
  }
}